Chnages on Project Show
